Agents/NotionAgent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "notion_agent"
version = "0.1.0"
description = "Agent for interacting with Notion workspaces"
authors = [
	{name = "Author"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
	"aiohttp",
	"python-dotenv",
	"pydantic>=1.10.8,<2.0.0",
	"pytest",
	"pytest-asyncio",
]

[project.optional-dependencies]
dev = [
	"black",
	"flake8",
	"mypy",
]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--import-mode=importlib"

[tool.setuptools]
packages = ["Agent", "operations", "launcher", "tests"] 